requirements.txt
# Basic
logzero
requests

# Selenium
selenium

# Timezone
pytz

# Slack
slackclient

# AWS / SSM
boto3
ssm-cache

# sqlite3
apsw